Version 2.3
Made the Constant text colour change more gradually as it goes over the 'danger' area. Previously the text colour went to jet-black suddenly as the background hit the dark boundary coming from from dark background. It looked silly, suddenly going black. Now the colour is a more nice dark-brown colour, not jet-black as it does this. Tuned-in the background 'down'-lighting of Constant at the dark boundary of the danger zone so it is exactly as dark as it needs to be, no more, and correspondingly tuned-in the in-the-danger-zone' background adjustment so that it brings out nicely the new non-jet-black text colour.