Version 2.3@1
Made the Constant look pretty at the lighter background area. Optimised as much as I could the visibility of Constant at lower background tones and the darker part of the clash-background area. Clash background is where background is very similar to text's colour and therefore is hardest for the eye to see.