Version 3.5
Made the transition into 'danger' area from dark->light| of the Constant a bit more gradual so as to maintain an overall illusion of progress. Previously the change had been abrupt (Constants went quite dark suddenly) so this illusion was lost. To do this modified the start and end times of the 'danger' area. Went ahead and added new easeArea global variable to allow the easing length to be controlled. This allowed me to go ahead and create longer easing for the Constant from dark->light as well, again smoothing the change from non-'danger' to 'danger'.