Version 4.6
Changed the speed at which background shifts from light to dark and vice-versa so that instead of going gradually from dark to light over the hour then back in the next hour, the background now does so in a half hour slots so the same as happened over two hours before now happens over one hour. The 'shaded backgrounds' added in 4.2 keep it interesting as on hour 1 the colour shades differently to what it does on hour 2, and again to what it does on hour 3. If you feel this is too sudden stick to script version 4.5 as this version does this more gradually over 2 hour slots.