Version 5.0
Realised that the background was not continuous on the second hour. Made the background fading continuous, so it looks like there's continuity. Hour 1 background fades smoothly between dark-blue to light-yellow. Hours 2 it goes smoothly from dark-green to a lighter green. Finally hour 3 the background goes smoothly from a dark-red to a light pinky colour. I feel that whilst these are not the prettiest shades ever, that they are at least doing the job. I need to try these out fully to get a feel for how they're working and hence get better shades, but for now they're working. (Look  smooth). Watch this space for better ideas.