Version 4.9
Made the visibility-protection of Identifier a bit less glary. Previously i was trying to use brightening of the background to improve visibility but it was too strong. This was my fault. I was trying to improve it. Apologies. I've now corrected this so that Identifier remains remain nice and easy on the eyes as the back- ground fades from dark-light and vice versa.