Version 6.4
The same as before but made the brightening work a bit more 'blended-in' with the background, so as the background lightens the Normal element kicks-in a bit stronger with its own lightening so that it doesn't get 'out-lightened' by the lighter background, while at the same time Normal doesn't lighten too strongly at darker background ranges to compensate for the fact that contrast is better anyway. Now Normal blends in very nicely throughout all background lightnesses.