Version 6.6
A bit of icing on the cake really, made the cursor turn a nice primary red at the highest background lightness. It's better than before as before it just turned a darker yellow and this was dull and unimaginative and didn't out. Now with the red the cursor looks nice and strong at a lighter background brightness.