Version 7.3
Tried my best to tune-up the Normal syntax element which is unique because it's background is the global background. Other elements can have their backgrounds "fixed-up" using various tricks to correct their visibility but Normal can't. Typically Normal is the keywords and most of the variable names in Javascript. Other languages rely on it to a lesser degree, but still can be quite heavily dependent on it. The better supported languages (like PHP) rely on it a lot less. Anyway, if you use Changing Colour script now Normal doesn't stand-out glaring too much but while still remaining visible at both bright and dark backgrounds. It's quite a subtle compromise between glare and visibility this one and i hope i got it right.