Version 7.4
Went carefuly over Statement as this was changing too agressively over the difficult visibility patch. Originally the visibility-tweaking functions were not as detailed so this reflected the limited options there were then. Now Statement fully uses 'ramped-brighten-ups' that as this term suggests brighten-up backgrounds to correct visibility problems in a smooth (not abrupt) manner.