Version 8.1
Corrected a slight glitch with the improved 'timer' routine that happened if the cursor was bang in the middle of the screen or the line above it. It got confused because what had been true previously had just changed. Now fixed. Did this by adding special cases of where to move 'safely' if the cursor is a) bang in the middle and b) directly *above* the middle. Did this using a 'middle' flag and an 'abovemiddle' flag. Ugly but it does the job. (Sorry, but I'm not Donald Knutz.)