Version 8.3
There was a bug with the 'timer' routine which made vim flash/beep if there was not enough lines above to temporarily move the cursor to (this causing the 'timer'.) This has now been fixed. If there are too few lines the timer will try to move left and right instead of up and down. There is a slight possibility that the editor will beep at you if you are on the first character or if there are no characters in the file, but hopefully that should not be the case long enough to bother. Also added a line set whichwrap=h,l to minimise this flashing. If you like using this script it is better to leave this on as it will minimise this flashing/beeping.