Version 8.4
Finally made the 'timer' well-behaved by detecting if the cursor was on the first column of a line in the case where there were too few lines to move up/down. The cursor now moves *right* if you're at column 1 (first column), *not* left as before. The cursor now moves left only if your cursor is *not* on column 1, rather than indiscriminately move left Anyway, beeping/ flashing should now almost go away. There is a case that I can't handle and that's when there is only one character in the file, in which case the editor beep/ flash once as it tries to do a 'timer' but there isn't any space to move 'right' to, but hopefully that shouldn't be too bad.