Version 9.3
Realised that i had left a slightly exposed area around the light area of the background. This is where background is light-ish but the Normal text is not near the dark-light exchange yet. fixed this by stretching the 'ease' area, or the 'ramping' zone of the normal text. This ramping zone now covers this previously exposed zone around the lighter background.