Version 9.1
Added a new 'dark dip' adjustable ramp to the 'Normal' syntax element, this allows a subtle incremental 'darkening' of the text as it nears the point where it will go into 'light' mode. This darkening helps keep the visibility up right up until the last minute. Maybe some fine-tuning needed but it's much better as it is than before. Before visibility just worsened around the point where Normal would switch from light text to dark text and vice-versa as the background darkened or lightened. This no longer is true.