Version 9.8
Made funky adjustments to ease the visual flow of Normal from dark-light background and vice versa by individually switching the RGB components up/down one at a time, e.g. first Red, and so on, as otherwise it's abrupt and causes dis-orientation. Also played with Statement and improved the visibility as at some of the dark-background settings it was looking a bit like a carrot-ey orange on top of mud - not very clear.