Version 10.0
On the 10.0 version, perhaps now the best version. Normal now blends more smoothly than ever at darker background ranges, and the background colour at the lighter ranges now peaks up to a pure white before coming down again to a tint which can be a shade of either light-green, cyan, rose. The contrast is great, the visibility is great, and the smoothness is great. If you haven't tried this before you'll really be amazed.