Version 11.0
Prettied-up Identifier and Statement. Still some non-vital glitches like sometimes the search/replace prompt is invisible, (fg==bg), folded lines text invisible (fg=bg) (which may not be so bad) and cursorline's colour is not always very distinct from general background (which may not be a problem if you don't use it.) But generally getting there.