Version 16.2
Removed the timer function from here and made it into a separate script.  This script needs that script in to work properly, so you should install that in your _vimrc if you have not already done so.  Download it from here: http://www.vim.org/scripts/script.php?script_id=3244