Version 16.4
Made improvements to visibility adjustments of Normal because it's very important it remains strong. The default 'color change' frequency is now ~15 seconds, not 1.5 mins.  I thought it was a bit more exciting and seems to work ok.  Have been testing it for months on four computers running XP. Made Cursor a bit more visible too.