Version 17.2@1
This is a quick mod that adds a bit of fizz to the editing, a \"standard\" funky scheme is picked but as soon as you stop editing the \"changing one\" to remind you of time passing.  Good for keeping your boss off your case.  You can change my choice in color scheme if you like, just find your plugin directory in vim, hint, type :echo $VIM$ or :echo $VIMRUNTIME$, on mac it&#039s in $HOME$/.vim/plugin, and change the bottom two lines of changingColor.vim.  You can change what \"event\" triggers this as well.