Version 19.3
Tweaked visibility of Normal, edge case.  It appears that due to a calibration issue I was having with my monitor I'd got this wrong previously.  I had thought that it was massively out but it was a "Power Saving feature of my EeePC that made it look hard to read. Not generally applicable in most cases.  Fixed to be similar to what it was in 19.0.