Version 20.0
After careful consideration have decided that some of the shades in my script are too difficult to see.  I was trying to have one smooth continuum between light and dark but the exact middle point is always making me strain to read no matter what I do, and this looks poor compared to other color schemes so I decided to create a ramped "jump over" the exact mid-point. You will still see constant changes but when it gets to a certain mid point it will jump about 20% of the murky shades to the next point after that.